Move distributions code to ManifoldDistributions
