+++ src/manifolds/Euclidean.jl
@@ -302,17 +302,6 @@ in this case, just the (Frobenius) norm of `X`.
 norm(::Euclidean, p, X) = norm(X)
 norm(::MetricManifold{<:Manifold,EuclideanMetric}, p, X) = norm(X)
 
-"""
-    normal_tvector_distribution(M::Euclidean, p, σ)
-
-Normal distribution in ambient space with standard deviation `σ`
-projected to tangent space at `p`.
-"""
-function normal_tvector_distribution(M::Euclidean{Tuple{N}}, p, σ) where {N}
-    d = Distributions.MvNormal(zero(p), σ)
-    return ProjectedFVectorDistribution(TangentBundleFibers(M), p, d, project_vector!, p)
-end
-
 @doc raw"""
     project(M::Euclidean, p)
 
@@ -333,19 +322,6 @@ space of `M` can be identified with all of `M`.
 project(::Euclidean, ::Any, ::Any)
 
 project!(M::Euclidean, Y, p, X) = copyto!(Y, X)
-
-"""
-    projected_distribution(M::Euclidean, d, [p])
-
-Wrap the standard distribution `d` into a manifold-valued distribution. Generated
-points will be of similar type to `p`. By default, the type is not changed.
-"""
-function projected_distribution(M::Euclidean, d, p)
-    return ProjectedPointDistribution(M, d, project!, p)
-end
-function projected_distribution(M::Euclidean, d)
-    return ProjectedPointDistribution(M, d, project!, rand(d))
-end
 
 """
     representation_size(M::Euclidean)
