+++ src/manifolds/MetricManifold.jl
@@ -126,11 +126,9 @@ decorator_transparent_dispatch(::typeof(mean), M::MetricManifold, args...) = Val
 decorator_transparent_dispatch(::typeof(mean!), M::MetricManifold, args...) = Val(:intransparent)
 decorator_transparent_dispatch(::typeof(median), M::MetricManifold, args...) = Val(:parent)
 decorator_transparent_dispatch(::typeof(median!), M::MetricManifold, args...) = Val(:intransparent)
-decorator_transparent_dispatch(::typeof(normal_tvector_distribution), M::MetricManifold, arge...) = Val(:intransparent)
 decorator_transparent_dispatch(::typeof(norm), M::MetricManifold, args...) = Val(:intransparent)
 decorator_transparent_dispatch(::typeof(project), M::MetricManifold, args...) = Val(:parent)
 decorator_transparent_dispatch(::typeof(project!), M::MetricManifold, args...) = Val(:intransparent)
-decorator_transparent_dispatch(::typeof(projected_distribution), M::MetricManifold, arge...) = Val(:intransparent)
 decorator_transparent_dispatch(::typeof(sharp), M::MetricManifold, args...) = Val(:parent)
 decorator_transparent_dispatch(::typeof(sharp!), M::MetricManifold, args...) = Val(:intransparent)
 decorator_transparent_dispatch(::typeof(retract), M::MetricManifold, args...) = Val(:parent)
