+++ test/metric.jl
@@ -53,8 +53,8 @@ Manifolds.get_basis(::BaseManifold{N},x,B::DefaultOrthonormalBasis) where {N} = 
 Manifolds.get_coordinates!(::BaseManifold, Y, p, X, ::DefaultOrthonormalBasis) = (Y .= X)
 Manifolds.get_vector!(::BaseManifold, Y, p, X, ::DefaultOrthonormalBasis) = (Y .= X)
 Manifolds.default_metric_dispatch(::BaseManifold, ::DefaultBaseManifoldMetric) = Val(true)
-ManifoldDistributions.projected_distribution(M::BaseManifold, d) = ProjectedPointDistribution(M, d, project!, rand(d))
-ManifoldDistributions.projected_distribution(M::BaseManifold, d, x) = ProjectedPointDistribution(M, d, project!, x)
+ManifoldDistributions.projected_distribution(M::BaseManifold, d) = ManifoldDistributions.ProjectedPointDistribution(M, d, project!, rand(d))
+ManifoldDistributions.projected_distribution(M::BaseManifold, d, x) = ManifoldDistributions.ProjectedPointDistribution(M, d, project!, x)
 Manifolds.mean!(::BaseManifold, y, x::AbstractVector, w::AbstractVector; kwargs...) = fill!(y, 1)
 Manifolds.median!(::BaseManifold, y, x::AbstractVector, w::AbstractVector; kwargs...) = fill!(y, 2)
 Manifolds.mean!(::MetricManifold{BaseManifold{N},BaseManifoldMetric{N}}, y, x::AbstractVector, w::AbstractVector; kwargs...) where {N} = fill!(y, 3)
