+++ test/rotations.jl
@@ -46,8 +46,8 @@ include("utils.jl")
             test_musical_isomorphisms = true,
             retraction_methods = retraction_methods,
             inverse_retraction_methods = inverse_retraction_methods,
-            point_distributions = [Manifolds.normal_rotation_distribution(M, pts[1], 1.0)],
-            tvector_distributions = [Manifolds.normal_tvector_distribution(M, pts[1], 1.0)],
+            point_distributions = [ManifoldDistributions.normal_rotation_distribution(M, pts[1], 1.0)],
+            tvector_distributions = [ManifoldDistributions.normal_tvector_distribution(M, pts[1], 1.0)],
             basis_types_to_from = basis_types)
 
         @testset "log edge cases" begin
@@ -69,10 +69,10 @@ include("utils.jl")
     end
 
     @testset "Distribution tests" begin
-        usd_mmatrix = Manifolds.normal_rotation_distribution(M, (@MMatrix [1.0 0.0; 0.0 1.0]), 1.0)
+        usd_mmatrix = ManifoldDistributions.normal_rotation_distribution(M, (@MMatrix [1.0 0.0; 0.0 1.0]), 1.0)
         @test isa(rand(usd_mmatrix), MMatrix)
 
-        gtsd_mvector = Manifolds.normal_tvector_distribution(M, (@MMatrix [1.0 0.0; 0.0 1.0]), 1.0)
+        gtsd_mvector = ManifoldDistributions.normal_tvector_distribution(M, (@MMatrix [1.0 0.0; 0.0 1.0]), 1.0)
         @test isa(rand(gtsd_mvector), MMatrix)
     end
 
@@ -80,8 +80,8 @@ include("utils.jl")
     for n ∈ (3, 4, 5)
         @testset "Rotations: SO($n)" begin
             SOn = Manifolds.Rotations(n)
-            ptd = Manifolds.normal_rotation_distribution(SOn, Matrix(1.0I, n, n), 1.0)
-            tvd = Manifolds.normal_tvector_distribution(SOn, Matrix(1.0I, n, n), 1.0)
+            ptd = ManifoldDistributions.normal_rotation_distribution(SOn, Matrix(1.0I, n, n), 1.0)
+            tvd = ManifoldDistributions.normal_tvector_distribution(SOn, Matrix(1.0I, n, n), 1.0)
             pts = [rand(ptd) for _ in 1:3]
             test_manifold(SOn, pts;
                 test_forward_diff = n==3,
